Split autoApprove into per-function modules
