+++ src/mcp/shellicar/consts.ts
@@ -1,0 +1,5 @@
+import type { ApproveRule } from './types';
+
+export const PRESET_RULES: Record<string, ApproveRule[]> = {
+  defaults: [{ program: '~/.claude/skills/*/scripts/*.sh' }],
+};
