+++ src/mcp/shellicar/escapeRegex.ts
@@ -1,0 +1,3 @@
+export function escapeRegex(s: string): string {
+  return s.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+}
