+++ src/mcp/shellicar/ruleMatchesArgs.ts
@@ -1,0 +1,5 @@
+import type { ApproveRule } from './types';
+
+export function ruleMatchesArgs(commandArgs: string[], rule: ApproveRule): boolean {
+  return !rule.args || rule.args.every((arg) => commandArgs.includes(arg));
+}
