+++ src/mcp/shellicar/types.ts
@@ -1,0 +1,9 @@
+export interface ApproveRule {
+  program: string;
+  args?: string[];
+}
+
+export interface ExecPermissions {
+  presets?: string[];
+  approve?: ApproveRule[];
+}
