+++ test/autoApprove.spec.ts
@@ -1,6 +1,6 @@
 import { ExecInputSchema } from '@shellicar/mcp-exec';
 import { describe, expect, it } from 'vitest';
-import { isExecAutoApproved } from '../src/mcp/shellicar/autoApprove';
+import { isExecAutoApproved } from '../src/mcp/shellicar/isExecAutoApproved';
 
 const HOME = process.env.HOME ?? '/home/testuser';
